refactor: move UserAvatar and update import paths across multiple components
